requirements.txt
Django==1.6.5
Markdown==2.5.1
Pillow==2.6.1
PyJWT==0.3.0
South==1.0.1
dj-database-url==0.3.0
dj-static==0.0.5
django-appconf==0.6
django-cache-url==0.6.0
django-classy-tags==0.4
django-configurations==0.8
django-crispy-forms==1.4.0
django-filter==0.8
django-redis-cache==0.13.0
django-social-auth==0.7.28
django-social-share==0.2.1
django-suit==0.2.11
django-twitter-tag==1.2
djangorestframework==2.4.3
gunicorn==18.0
hiredis==0.1.4
httplib2==0.9
newrelic==2.38.0.31
oauth2==1.5.211
oauthlib==0.7.1
psycopg2==2.5.2
pystache==0.5.4
python-openid==2.2.5
python-social-auth==0.2.1
redis==2.10.1
requests==2.4.3
requests-oauthlib==0.4.2
six==1.2.0
sorl-thumbnail==11.12
static==1.0.2
twitter==1.9.1
wsgiref==0.1.2